updated the functions to call the correct rest end point
