+++ src/main/java/dev/appsody/starter/health/StarterLivenessCheck.java
@@ -1,10 +1,10 @@
 package dev.appsody.starter.health;
 
-import javax.enterprise.context.ApplicationScoped;
-
 import org.eclipse.microprofile.health.HealthCheck;
 import org.eclipse.microprofile.health.HealthCheckResponse;
 import org.eclipse.microprofile.health.Liveness;
+
+import javax.enterprise.context.ApplicationScoped;
 
 @Liveness
 @ApplicationScoped
@@ -15,11 +15,11 @@ public class StarterLivenessCheck implements HealthCheck {
 
         return true;
     }
-	
+
     @Override
     public HealthCheckResponse call() {
         boolean up = isAlive();
         return HealthCheckResponse.named(this.getClass().getSimpleName()).state(up).build();
     }
-    
+
 }
