+++ src/main/resources/microprofile-config.properties
@@ -1,0 +1,4 @@
+jdbcURL=the_url_to_the_db
+dbuser=db_username
+dbpassword=db_password
+
